refactor: isolate canonical linear relation.
